bpaf refactor: validate more data inside the parser
One of the main ideas of bpaf is to perform as much of validation within
the parser as possible so if you have the config back out - it is valid.

This commit introduces Protocol enum, adds FromStr instance and uses it
instead if passing a string directly.

This refactor caught a minor problem in qr.rs: it should not be possible
for encryption to contain "nopass" so a check was redundant